--- a/rust/arrow/src/array/transform/utils.rs
+++ b/rust/arrow/src/array/transform/utils.rs
@@ -23,9 +23,7 @@ use crate::{
 #[inline]
 pub(super) fn reserve_for_bits(buffer: &mut MutableBuffer, len: usize) {
     let needed_bytes = bit_util::ceil(len, 8);
-    if buffer.len() < needed_bytes {
-        buffer.extend(needed_bytes - buffer.len());
-    }
+    buffer.resize(needed_bytes, 0);
 }
 
 /// sets all bits on `write_data` on the range `[offset_write..offset_write+len]` to be equal to the
@@ -53,7 +51,7 @@ pub(super) fn extend_offsets<T: OffsetSizeTrait>(
     mut last_offset: T,
     offsets: &[T],
 ) {
-    buffer.reserve(buffer.len() + offsets.len() * std::mem::size_of::<T>());
+    buffer.reserve(offsets.len() * std::mem::size_of::<T>());
     offsets.windows(2).for_each(|offsets| {
         // compute the new offset
         let length = offsets[1] - offsets[0];