--- a/rust/arrow/src/compute/kernels/comparison.rs
+++ b/rust/arrow/src/compute/kernels/comparison.rs
@@ -48,9 +48,8 @@ macro_rules! compare_op {
             combine_option_bitmap($left.data_ref(), $right.data_ref(), $left.len())?;
 
         let byte_capacity = bit_util::ceil($left.len(), 8);
-        let actual_capacity = bit_util::round_upto_multiple_of_64(byte_capacity);
-        let mut buffer = MutableBuffer::new(actual_capacity);
-        buffer.resize(byte_capacity);
+        let mut buffer = MutableBuffer::new(byte_capacity);
+        buffer.resize(byte_capacity, 0);
         let data = buffer.as_mut_ptr();
 
         for i in 0..$left.len() {
@@ -81,9 +80,8 @@ macro_rules! compare_op_scalar {
         let null_bit_buffer = $left.data().null_buffer().cloned();
 
         let byte_capacity = bit_util::ceil($left.len(), 8);
-        let actual_capacity = bit_util::round_upto_multiple_of_64(byte_capacity);
-        let mut buffer = MutableBuffer::new(actual_capacity);
-        buffer.resize(byte_capacity);
+        let mut buffer = MutableBuffer::new(byte_capacity);
+        buffer.resize(byte_capacity, 0);
         let data = buffer.as_mut_ptr();
 
         for i in 0..$left.len() {