--- a/rust/arrow/src/compute/kernels/take.rs
+++ b/rust/arrow/src/compute/kernels/take.rs
@@ -272,7 +272,7 @@ where
     let data_len = indices.len();
 
     let mut buffer = MutableBuffer::new(data_len * std::mem::size_of::<T::Native>());
-    buffer.resize(data_len * std::mem::size_of::<T::Native>());
+    buffer.resize(data_len * std::mem::size_of::<T::Native>(), 0);
     let data = buffer.typed_data_mut();
 
     let nulls;
@@ -417,7 +417,7 @@ where
 
     let bytes_offset = (data_len + 1) * std::mem::size_of::<OffsetSize>();
     let mut offsets_buffer = MutableBuffer::new(bytes_offset);
-    offsets_buffer.resize(bytes_offset);
+    offsets_buffer.resize(bytes_offset, 0);
 
     let offsets = offsets_buffer.typed_data_mut();
     let mut values = Vec::with_capacity(bytes_offset);