@jorgecarleitao 2021-01-02
@nevi-me , there is a failing test in parquet that I am unable to fix. I think that we may be creating bitmap buffers that are too large.

My current hypothesis is that the buffer `old_bitmap` at `consume_bitmap_buffer` is not being resized in the same way the other buffer in `consume_rep_levels` is, which is causing it to be too large. Do you agree with this (before I proceed to patch it)?