@Dandandan 2021-01-19
@jorgecarleitao I think this PR broke master:
```
   --> datafusion/src/physical_plan/parquet.rs:712:25
    |
712 |             data_buffer.resize(data_buffer.len() + data_size);
    |                         ^^^^^^ ----------------------------- supplied 1 argument
    |                         |
    |                         expected 2 arguments
    |
note: associated function defined here
   --> rust/arrow/src/buffer.rs:833:12
    |
833 |     pub fn resize(&mut self, new_len: usize, value: u8) {
```